removed rocksdb in favour of inmemorydedup
