--- dependencies/maven/artifacts/org/rocksdb/BUILD
@@ -1,11 +1,0 @@
-java_library(
-    name = "rocksdbjni",
-    exports = [
-        "//external:jar/org/rocksdb/rocksdbjni"
-    ],
-    visibility = [
-        "//visibility:public"
-    ]
-)
-
-
