+++ dependencies/maven/dependencies.bzl
@@ -700,7 +700,6 @@ def list_dependencies():
     {"artifact": "org.ow2.asm:asm:5.0.4", "lang": "java", "sha1": "0da08b8cce7bbf903602a25a3a163ae252435795", "sha256": "896618ed8ae62702521a78bc7be42b7c491a08e6920a15f89a3ecdec31e9a220", "repository": "https://repo.maven.apache.org/maven2/", "url": "https://repo.maven.apache.org/maven2/org/ow2/asm/asm/5.0.4/asm-5.0.4.jar", "source": {"sha1": "112ff54474f1f04ccf1384c92e39fdc566f0bb5e", "sha256": "7ba89bc14669d86c1c0dc6abaeb74a87715089f3b904cc2016969e8737d70707", "repository": "https://repo.maven.apache.org/maven2/", "url": "https://repo.maven.apache.org/maven2/org/ow2/asm/asm/5.0.4/asm-5.0.4-sources.jar"} , "name": "org-ow2-asm-asm", "actual": "@org-ow2-asm-asm//jar", "bind": "jar/org/ow2/asm/asm"},
     {"artifact": "org.reflections:reflections:0.9.9-RC1", "lang": "java", "sha1": "b78b545f452a6b7d4fab2641dd0b0147a0f4fd5e", "sha256": "8be6ea8a7d790056c3492c2c9419985834b258345c81764984b4eb2a86c28492", "repository": "https://repo.maven.apache.org/maven2/", "url": "https://repo.maven.apache.org/maven2/org/reflections/reflections/0.9.9-RC1/reflections-0.9.9-RC1.jar", "source": {"sha1": "654dcaafdb84f03858bb69bbe031f0325bb49065", "sha256": "e495a0cde2c1348d9ee126573bb0bcd6cc87999ae8056cbbf0abe6349576aaf9", "repository": "https://repo.maven.apache.org/maven2/", "url": "https://repo.maven.apache.org/maven2/org/reflections/reflections/0.9.9-RC1/reflections-0.9.9-RC1-sources.jar"} , "name": "org-reflections-reflections", "actual": "@org-reflections-reflections//jar", "bind": "jar/org/reflections/reflections"},
     {"artifact": "org.roaringbitmap:RoaringBitmap:0.5.11", "lang": "java", "sha1": "e6b04760ea1896fc36beea4f11b8649481bf5af7", "sha256": "bad2fcf146d4a41cb188a28a4216f15cc470cd2fb6417a7d0005e7f8221ff312", "repository": "https://repo.maven.apache.org/maven2/", "url": "https://repo.maven.apache.org/maven2/org/roaringbitmap/RoaringBitmap/0.5.11/RoaringBitmap-0.5.11.jar", "source": {"sha1": "dd4dc1e5ed4d98ff8421e0e26f9822873c455e34", "sha256": "15767ab00bd7b50f8711003dd544b18d446a8ef9cdba06b6ad441cb8d2fe59b3", "repository": "https://repo.maven.apache.org/maven2/", "url": "https://repo.maven.apache.org/maven2/org/roaringbitmap/RoaringBitmap/0.5.11/RoaringBitmap-0.5.11-sources.jar"} , "name": "org-roaringbitmap-RoaringBitmap", "actual": "@org-roaringbitmap-RoaringBitmap//jar", "bind": "jar/org/roaringbitmap/RoaringBitmap"},
-    {"artifact": "org.rocksdb:rocksdbjni:5.14.2", "lang": "java", "sha1": "a6087318fab540ba0b4c6ff68475ffbedc0b3d10", "sha256": "34b7e45d18bca957e38cad0c3269dd36c9df81313f2ff41171ec2a96a3736c7f", "repository": "https://repo.maven.apache.org/maven2/", "url": "https://repo.maven.apache.org/maven2/org/rocksdb/rocksdbjni/5.14.2/rocksdbjni-5.14.2.jar", "source": {"sha1": "ac0184f4618db881be0a9aeac98535b13f36a967", "sha256": "8fdf11a3b7b19201459cbac18d7a693331a992c159f042ab6ece7e843f6ff59b", "repository": "https://repo.maven.apache.org/maven2/", "url": "https://repo.maven.apache.org/maven2/org/rocksdb/rocksdbjni/5.14.2/rocksdbjni-5.14.2-sources.jar"} , "name": "org-rocksdb-rocksdbjni", "actual": "@org-rocksdb-rocksdbjni//jar", "bind": "jar/org/rocksdb/rocksdbjni"},
     {"artifact": "org.scala-lang.modules:scala-parser-combinators_2.11:1.0.1", "lang": "java", "sha1": "f05d7345bf5a58924f2837c6c1f4d73a938e1ff0", "sha256": "19495ce701fd9ba3c499c137e3ad5364acee8f87a01ef99e912f995a23fd4cb1", "repository": "https://repo.maven.apache.org/maven2/", "url": "https://repo.maven.apache.org/maven2/org/scala-lang/modules/scala-parser-combinators_2.11/1.0.1/scala-parser-combinators_2.11-1.0.1.jar", "source": {"sha1": "34d013c02d0b73794ba2911552896dd9c00f34c3", "sha256": "d8543ae4bed3db45361f9dfd73cb1343d6969aa371831d0e836ab99a11276d85", "repository": "https://repo.maven.apache.org/maven2/", "url": "https://repo.maven.apache.org/maven2/org/scala-lang/modules/scala-parser-combinators_2.11/1.0.1/scala-parser-combinators_2.11-1.0.1-sources.jar"} , "name": "org-scala-lang-modules-scala-parser-combinators_2-11", "actual": "@org-scala-lang-modules-scala-parser-combinators_2-11//jar", "bind": "jar/org/scala-lang/modules/scala-parser-combinators-2-11"},
 # duplicates in org.scala-lang.modules:scala-xml_2.11 promoted to 1.0.5
 # - org.apache.tinkerpop:spark-gremlin:3.3.3 wanted version 1.0.5
